+++ Produlytics/persistence/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>it.deltax.produlytics</groupId>
     <artifactId>persistence</artifactId>
-    <version>0.0.2</version>
+    <version>0.0.1</version>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
