+++ Produlytics/persistence/src/main/java/it/deltax/produlytics/persistence/CharacteristicEntityId.java
@@ -17,8 +17,7 @@ public class CharacteristicEntityId implements Serializable {
 
 	protected CharacteristicEntityId() {}
 
-	public CharacteristicEntityId(Integer id, Integer deviceId) {
-		this.id = id;
+	public CharacteristicEntityId(Integer deviceId) {
 		this.deviceId = deviceId;
 	}
