+++ Produlytics/ui/back-end/pom.xml
@@ -42,7 +42,7 @@
 		<dependency>
 			<groupId>it.deltax.produlytics</groupId>
 			<artifactId>persistence</artifactId>
-			<version>0.0.2</version>
+			<version>0.0.1</version>
 		</dependency>
 	</dependencies>
