--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/detection/business/domain/DetectionLight.java
@@ -1,8 +1,0 @@
-package it.deltax.produlytics.uibackend.common.detection.business.domain;
-
-public record DetectionLight(
-	double value,
-	long createdAtUtc,
-	boolean anomalous
-) {
-}
