--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/devices/DevicesConfiguration.java
@@ -1,2 +1,0 @@
-package it.deltax.produlytics.uibackend.common.devices;public class DevicesConfiguration {
-}
