--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/machine/business/domain/CharacteristicDisplayInfo.java
@@ -1,7 +1,0 @@
-package it.deltax.produlytics.uibackend.common.machine.business.domain;
-
-public record CharacteristicDisplayInfo(
-	MachineLight machine,
-	Characteristic characteristic
-) {
-}
