--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/machine/business/domain/CharacteristicLight.java
@@ -1,9 +1,0 @@
-package it.deltax.produlytics.uibackend.common.machine.business.domain;
-
-public record CharacteristicLight(
-	long code,
-	String name,
-	long machine
-) {
-}
-
