--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/machine/business/domain/MachineLight.java
@@ -1,4 +1,0 @@
-package it.deltax.produlytics.uibackend.common.machine.business.domain;
-
-// business class for light lookups
-public record MachineLight(long id, String name) { }
