--- Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/common/machine/business/ports/in/FindCharacteristicInfoUseCase.java
@@ -1,9 +1,0 @@
-package it.deltax.produlytics.uibackend.common.machine.business.ports.in;
-
-import it.deltax.produlytics.uibackend.common.machine.business.domain.CharacteristicDisplayInfo;
-
-import java.util.Optional;
-
-public interface FindCharacteristicInfoUseCase {
-    Optional<CharacteristicDisplayInfo> find(int machineId, int id);
-}
