+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/detection/business/domain/DetectionLight.java
@@ -1,0 +1,8 @@
+package it.deltax.produlytics.uibackend.detection.business.domain;
+
+public record DetectionLight(
+	double value,
+	long createdAtUtc,
+	boolean anomalous
+) {
+}
