+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/devices/DevicesConfiguration.java
@@ -1,0 +1,2 @@
+package it.deltax.produlytics.uibackend.devices;public class DevicesConfiguration {
+}
