+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/domain/Characteristic.java
@@ -1,0 +1,14 @@
+package it.deltax.produlytics.uibackend.machine.business.domain;
+
+import java.math.BigDecimal;
+import java.util.Optional;
+
+public record Characteristic(
+	long id,
+	String name,
+	long machine,
+	Optional<Double> lowerLimit,
+	Optional<Double> upperLimit,
+	Optional<Double> average
+) {
+}
