+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/domain/CharacteristicDisplayInfo.java
@@ -1,0 +1,7 @@
+package it.deltax.produlytics.uibackend.machine.business.domain;
+
+public record CharacteristicDisplayInfo(
+	MachineLight machine,
+	Characteristic characteristic
+) {
+}
