+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/domain/CharacteristicLight.java
@@ -1,0 +1,9 @@
+package it.deltax.produlytics.uibackend.machine.business.domain;
+
+public record CharacteristicLight(
+	long code,
+	String name,
+	long machine
+) {
+}
+
