+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/domain/MachineLight.java
@@ -1,0 +1,4 @@
+package it.deltax.produlytics.uibackend.machine.business.domain;
+
+// business class for light lookups
+public record MachineLight(long id, String name) { }
