+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/ports/in/FindCharacteristicInfoUseCase.java
@@ -1,0 +1,9 @@
+package it.deltax.produlytics.uibackend.machine.business.ports.in;
+
+import it.deltax.produlytics.uibackend.machine.business.domain.CharacteristicDisplayInfo;
+
+import java.util.Optional;
+
+public interface FindCharacteristicInfoUseCase {
+    Optional<CharacteristicDisplayInfo> find(int machineId, int id);
+}
