+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/ports/in/ListAllMachinesUseCase.java
@@ -1,0 +1,9 @@
+package it.deltax.produlytics.uibackend.machine.business.ports.in;
+
+import it.deltax.produlytics.uibackend.machine.business.domain.MachineLight;
+
+import java.util.List;
+
+public interface ListAllMachinesUseCase {
+    List<MachineLight> listAll();
+}
