+++ Produlytics/ui/back-end/src/main/java/it/deltax/produlytics/uibackend/machine/business/ports/in/ListCharacteristicsByMachineUseCase.java
@@ -1,0 +1,9 @@
+package it.deltax.produlytics.uibackend.machine.business.ports.in;
+
+import it.deltax.produlytics.uibackend.machine.business.domain.CharacteristicLight;
+
+import java.util.List;
+
+public interface ListCharacteristicsByMachineUseCase {
+    List<CharacteristicLight> listByMachine(int machineId);
+}
